docs:강의자료 정리

+++ DB/db/02-02-many-to-one-relationships/articles/views.py
@@ -31,7 +31,9 @@ def create(request):
     if request.method == 'POST':
         form = ArticleForm(request.POST)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -44,8 +46,8 @@ def create(request):
 @login_required
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
-
-    article.delete()
+    if request.user == article.user:
+        article.delete()
 
     return redirect('articles:index')
 
@@ -53,16 +55,19 @@ def delete(request, pk):
 @login_required
 def update(request, pk):
     article = Article.objects.get(pk=pk)
+    # 현재 수정을 요청하는 유저와 게시글의 작성자가 같은지 확인
+    if request.user == article.user:
+        if request.method == 'POST':
+            # 기존의 데이터를 받기 때문에 (instance=article) user 정보 안들어가도 됨 
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article.pk)
 
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:detail', article.pk)
-
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
-
+        return redirect('articles:index')
     context = {
         'article': article,
         'form': form,
@@ -83,6 +88,7 @@ def comments_create(request, pk):
         comment = comment_form.save(commit=False)
         # 외래 키 데이터를 할당
         comment.article = article
+        comment.user = request.user
         comment.save()
         return redirect('articles:detail', article.pk)
     context = {
@@ -96,7 +102,8 @@ def comments_create(request, pk):
 def comments_delete(request, article_pk, comment_pk):
     # 삭제할 댓글 조회
     comment = Comment.objects.get(pk=comment_pk)
-    # 댓글 삭제
-    comment.delete()
+    if request.user == comment.user:
+        # 댓글 삭제
+        comment.delete()
     # 삭제 후 게시글 상세 페이지로 리다이렉트
     return redirect('articles:detail', article_pk)
